requirements.txt
pyside6==6.7.2
pillow==10.4.0
opencv-python==4.10.0.84